평행한 직선 충돌 처리 필요

+++ Objects/Background_Yeti.cpp
@@ -5,7 +5,7 @@ Background_Yeti::Background_Yeti()
 {
 	mapSprite = new Sprite(Textures + L"/TianSouls/yeti_map.png", Shaders + L"009_Sprite.fx",
 		1704, 983, 2036, 1349);
-	mapSprite->SetAbsoluteScale(Width, Height);
+	mapSprite->SetAbsoluteScale((float)Width, (float)Height);
 	//mapSprite->Position(Width * 0.5f, Height * 0.5f);
 	mapSprite->Position(0, 0);
 	mapSprite->Rotation(0, 0, 0);
